+++ file.cpp
@@ -17,15 +17,19 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <exception>
 #include <fstream>
+#include <json/value.h>
 #include <string>
+#include <vector>
+#include <nlohmann/json.hpp>
 
 #include "file.hpp"
 
 // write to the file
-void write(const arguments& args)
+void write(const arguments& args, const Json::Value& config)
 {
-	std::ofstream file{open_file(args.fileName)};
+	std::ofstream file{add_extension(args.fileName, ".tex")};
 
 	file << "\\documentclass{" << args.docClass << "}\n\n"
 	<< "\\usepackage[T1]{fontenc}\n\n"
@@ -37,8 +41,9 @@ void write(const arguments& args)
 		file << '\n';
 		for (const std::string& package : args.packages)
 		{
-			file << "\\usepackage{" << package << "}\n";
+			file << add_options(config, package);
 		}
+		file << '\n';
 	}
 
 	if (args.bibsources.size() > 0)
@@ -53,20 +58,46 @@ void write(const arguments& args)
 	file << (args.title != ""  ? "\\title{" + args.title + "}\n" : "")
 	<< (args.author != "" ? "\\author{" + args.author + "}\n" : "")
 	<< (args.date != "" ? "\\date{" + args.date + "}\n" : "")
-	<< "\n\\begin{document}\n\t\n\\end{document}\n";
-}
-
-// open the output file for writing
-std::ofstream open_file(const std::string& name)
-{
-	// append the ".tex" extension if necesary
-	std::ofstream file{add_extension(name, ".tex")};
-
-	return file;
+	<< "\n\\begin{document}\n" << (args.author != "" || args.title != "" ? "\t\\maketitle\n" : "" )
+	<< "\t\n\\end{document}\n";
 }
 
 // add the filextension if necessary
 std::string add_extension(const std::string& fileName, const std::string&  extension)
 {
 	return (fileName.find(extension) == std::string::npos ? fileName + extension: fileName);
+}
+
+std::string add_options(const Json::Value& config, const std::string& package)
+{
+	std::vector<std::string> options{};
+	if (config["options"][package])
+	{
+		for (auto& option : config["options"][package])
+		{
+			options.push_back(option.asString());
+		}
+	}
+
+	if (options.size() > 0)
+	{
+		std::string command{"\\usepackage["};
+		for (unsigned long i{0}, l{options.size()}; i < l; ++i)
+		{
+			if (i != l - 1)
+			{
+				command += options[i] + ", ";
+			}
+			else
+			{
+				command += options[i] + "]{" + package + "}\n";
+			}
+		}
+		return command;
+	}
+	else
+	{
+		std::string command{"\\usepackage{" + package + "}\n"};
+		return command;
+	}
 }